Refactor kernel panic handling and enhance diagnostics
- Updated panic handler to include detailed CPU, thread, and memory diagnostics.
- Added register snapshots for x86-64 and aarch64 architectures during panic.
- Improved heap allocation failure handling with diagnostics.
- Modified spinlock implementation to track timeout hits for better debugging.
- Enhanced user pointer validation to check for mapped pages in memory.
- Updated syscall handlers to include new flags for thread information.
- Changed process ID generation to use atomic operations for thread safety.
- Implemented task grouping in the task manager for better process visualization.
- Added support for displaying child threads under their parent process in the UI.
- Improved memory and CPU usage reporting for grouped processes.

+++ kernel/src/net/tcp/connect.rs
@@ -183,7 +183,7 @@ pub fn accept(listener_id: u32, timeout_ticks: u32) -> (u32, Ipv4Addr, u16) {
             // Find a ready connection spawned by this listener
             for i in 0..table.len() {
                 let ready = table[i].as_ref().map(|tcb| {
-                    tcb.parent_listener == Some(lid as u8)
+                    tcb.parent_listener == Some(lid as u16)
                         && tcb.state == TcpState::Established
                         && !tcb.accepted
                 }).unwrap_or(false);
@@ -248,7 +248,7 @@ pub fn close_listener(socket_id: u32) -> u32 {
     // Clean up pending connections
     for i in 0..table.len() {
         let is_pending = table[i].as_ref().map(|tcb| {
-            tcb.parent_listener == Some(id as u8) && !tcb.accepted
+            tcb.parent_listener == Some(id as u16) && !tcb.accepted
         }).unwrap_or(false);
         if is_pending {
             table[i] = None;
